refactor: Organize hooks & automatically scroll to the top of the viewport when routes change
