--- client/src/hooks.js
@@ -1,11 +1,0 @@
-import { useContext } from 'react';
-import { UserContext } from 'providers/UserProvider';
-import { ModalContext } from 'providers/ModalProvider';
-
-export const useCurrentUser = () => {
-  return useContext(UserContext);
-};
-
-export const useModal = () => {
-  return useContext(ModalContext);
-};
