+++ client/src/hooks/useCurrentUser.js
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { UserContext } from 'providers/UserProvider';
+
+const useCurrentUser = () => {
+  return useContext(UserContext);
+};
+
+export default useCurrentUser;
