+++ client/src/hooks/useModal.js
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { ModalContext } from 'providers/ModalProvider';
+
+const useModal = () => {
+  return useContext(ModalContext);
+};
+
+export default useModal;
